requirements.txt
jinja2
paramiko
pyyaml
PySide6
QtAwesome
